WIP Frontend interval values from ppp and resolution to points
