+++ ui/src/dashboards/containers/DashboardPage.js
@@ -26,6 +26,7 @@ import {
   templateControlBarVisibilityToggled as templateControlBarVisibilityToggledAction,
 } from 'shared/actions/app'
 import {presentationButtonDispatcher} from 'shared/dispatchers'
+import {interval} from 'src/shared/constants'
 
 const FORMAT_INFLUXQL = 'influxql'
 const defaultTimeRange = {
@@ -278,25 +279,6 @@ class DashboardPage extends Component {
         {
           value: up || 'now()',
           type: upperType,
-          selected: true,
-        },
-      ],
-    }
-
-    const interval = {
-      id: 'interval',
-      type: 'autoGroupBy',
-      tempVar: ':interval:',
-      label: 'automatically determine the best group by time',
-      values: [
-        {
-          value: '1000', // pixels
-          type: 'resolution',
-          selected: true,
-        },
-        {
-          value: '3',
-          type: 'pointsPerPixel',
           selected: true,
         },
       ],
