+++ ui/src/data_explorer/components/Table.js
@@ -7,7 +7,7 @@ import {Table, Column, Cell} from 'fixed-data-table'
 import Dropdown from 'shared/components/Dropdown'
 import CustomCell from 'src/data_explorer/components/CustomCell'
 import TabItem from 'src/data_explorer/components/TableTabItem'
-import {TEMPLATES} from 'src/data_explorer/constants'
+import {TEMPLATES} from 'src/shared/constants'
 
 import {fetchTimeSeriesAsync} from 'shared/actions/timeSeries'
