+++ ui/src/data_explorer/components/VisHeader.js
@@ -5,7 +5,7 @@ import _ from 'lodash'
 import {fetchTimeSeriesAsync} from 'shared/actions/timeSeries'
 import {resultsToCSV} from 'src/shared/parsing/resultsToCSV.js'
 import download from 'src/external/download.js'
-import {TEMPLATES} from 'src/data_explorer/constants'
+import {TEMPLATES} from 'src/shared/constants'
 
 const getCSV = (query, errorThrown) => async () => {
   try {
