+++ ui/src/data_explorer/containers/DataExplorer.js
@@ -6,16 +6,16 @@ import queryString from 'query-string'
 
 import _ from 'lodash'
 
-import QueryMaker from '../components/QueryMaker'
-import Visualization from '../components/Visualization'
+import QueryMaker from 'src/data_explorer/components/QueryMaker'
+import Visualization from 'src/data_explorer/components/Visualization'
 import WriteDataForm from 'src/data_explorer/components/WriteDataForm'
-import Header from '../containers/Header'
-import ResizeContainer from 'shared/components/ResizeContainer'
-import OverlayTechnologies from 'shared/components/OverlayTechnologies'
+import Header from 'src/data_explorer/containers/Header'
+import ResizeContainer from 'src/shared/components/ResizeContainer'
+import OverlayTechnologies from 'src/shared/components/OverlayTechnologies'
 import ManualRefresh from 'src/shared/components/ManualRefresh'
 
-import {VIS_VIEWS, AUTO_GROUP_BY} from 'shared/constants'
-import {MINIMUM_HEIGHTS, INITIAL_HEIGHTS, TEMPLATES} from '../constants'
+import {VIS_VIEWS, AUTO_GROUP_BY, TEMPLATES} from 'src/shared/constants'
+import {MINIMUM_HEIGHTS, INITIAL_HEIGHTS} from 'src/data_explorer/constants'
 import {errorThrown} from 'shared/actions/errors'
 import {setAutoRefresh} from 'shared/actions/app'
 import * as dataExplorerActionCreators from 'src/data_explorer/actions/view'
